vk_rasterizer: Move vertex and index buffer bind to rasterizer
Feels cleaner in general and wont have to pass barriers

+++ src/shader_recompiler/backend/spirv/spirv_emit_context.cpp
@@ -1172,8 +1172,9 @@ Id EmitContext::DefineUfloatM5ToFloat32(u32 mantissa_bits, const std::string_vie
 }
 
 Id EmitContext::DefineGetBdaPointer() {
-    const auto caching_pagebits{Constant(U64, profile.sparse_page_shift)};
-    const auto caching_pagemask{Constant(U64, (1u << profile.sparse_page_shift) - 1)};
+    const auto caching_pagebits{Constant(U64, static_cast<u64>(profile.sparse_page_shift))};
+    const auto caching_pagemask{
+        Constant(U64, static_cast<u64>((1u << profile.sparse_page_shift) - 1))};
 
     const auto func_type{TypeFunction(U64, U64)};
     const auto func{OpFunction(U64, spv::FunctionControlMask::MaskNone, func_type)};
